// ClientApp/src/app/models/personnel-info.ts
export class PersonnelInfo {
    // PRSNEL_ID: number;
    // PRSNEL_LAST_NM: string;
    // PRSNEL_FIRST_NM: string;
    // EMAIL_ADDR_TXT: string;
    prsnel_id: string;
    LAST_NAME: string;
    FIRST_NAME: string;
    EMAIL_ADDR_TXT: string;
    DISPLAY_NAME: string;
    LAST_LOGGED_IN: string;
}